agents: break the tool-permissions/capabilities import cycle that deadlocked the bundled backend
gateTools moves to capabilities/gate-tools.ts so the import edge stays
one-way (capabilities -> tool-permissions, never back). The reverse edge
turned the esbuild bundle's async module wrappers into a cycle —
init_tool_permissions awaiting init_adapters awaiting
init_tool_permissions — and server.mjs died on an unsettled top-level
await before serving /health (the Quality Gate docker leg's failure).
backend:smoke now passes.
